bug fixes, main + admin pages ready, search and role forms ready

+++ src/main/java/by/gurinovich/ZapolZachetSpring/services/LabaService.java
@@ -1,25 +1,26 @@
 package by.gurinovich.ZapolZachetSpring.services;
 
 import by.gurinovich.ZapolZachetSpring.models.Laba;
+import by.gurinovich.ZapolZachetSpring.models.Student;
 import by.gurinovich.ZapolZachetSpring.models.Subject;
 import by.gurinovich.ZapolZachetSpring.repositories.LabaRepository;
 import by.gurinovich.ZapolZachetSpring.repositories.ZachetRepository;
-import org.springframework.beans.factory.annotation.Autowired;
+import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import java.util.Optional;
+import java.util.ArrayList;
+import java.util.List;
 
 @Service
 @Transactional(readOnly = true)
+@RequiredArgsConstructor
 public class LabaService {
     private final LabaRepository labaRepository;
     private final ZachetRepository zachetRepository;
-    @Autowired
-    public LabaService(LabaRepository labaRepository, ZachetRepository zachetRepository) {
-        this.labaRepository = labaRepository;
-        this.zachetRepository = zachetRepository;
-    }
+    private final SubjectService subjectService;
+    private final ZachetService zachetService;
+
 
     public Laba getById(Long id){
         return labaRepository.findById(id).orElse(null);
@@ -29,22 +30,45 @@ public class LabaService {
         return labaRepository.findByNumberAndSubject(number, subject).orElse(null);
     }
 
+    public List<Laba> getAllUndoneByStudentAndSubject(Student student, Subject subject){
+        List<Laba> result = new ArrayList<>();
+        student.getZachety().forEach(el -> result.addAll(labaRepository.findAllBySubjectAndZachetsNotContaining(subject, el)));
+        result.forEach(System.out::println);
+        return result;
+    }
+
     @Transactional
-    public Laba save(Integer number, String title, Subject subject){
-        Optional<Laba> temp = labaRepository.findByNumberAndSubject(number, subject);
-        return temp.orElseGet(() ->
-                labaRepository.save(Laba.builder()
-                                .title(title)
-                                .number(number)
-                                .subject(subject)
-                        .build()));
+    public Laba save(Laba laba){
+        Laba newLaba = addLabaToSubjects(laba);
+        addZachetsToStudents(laba);
+        return newLaba;
+    }
+
+    @Transactional
+    public Laba addLabaToSubjects(Laba laba){
+        Laba newLaba = labaRepository.save(laba);
+        Subject subject = laba.getSubject();
+        List<Laba> labas = subject.getLabas();
+        labas.add(newLaba);
+        subject.setLabas(labas);
+        subjectService.save(subject);
+        return newLaba;
+    }
+
+    @Transactional
+    public void addZachetsToStudents(Laba laba){
+        laba.getSubject().getGroups().forEach(group ->
+                group.getStudents().forEach(student ->
+                        zachetService.addZachetToStudent(student, laba)));
     }
 
     @Transactional
     public void deleteById(Long labaId){
-        Laba laba =  labaRepository.findById(labaId).orElse(null);
-        zachetRepository.deleteAllByLaba(laba);
-        labaRepository.delete(laba);
+        Laba laba =  getById(labaId);
+        if (laba != null) {
+            zachetRepository.deleteAllByLaba(laba);
+            labaRepository.delete(laba);
+        }
     }
 
 }
